backend/requirements.txt
asgiref==3.9.1
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework-api-key==3.1.0
ftfy==6.3.1
gunicorn==23.0.0
langcodes==3.5.0
language_data==1.3.0
locate==1.1.1
marisa-trie==1.2.1
msgpack==1.1.1
packaging==25.0
pillow==11.3.0
psycopg2==2.9.10
python-dotenv==1.1.1
regex==2025.7.34
setuptools==80.9.0
sqlparse==0.5.3
wcwidth==0.2.13
wordfreq==3.1.1
